add landingpage

+++ app/Http/Controllers/WebinarController.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Mentor;
 use App\Models\Webinar;
 use App\Models\WebinarCategory;
+use App\Models\WebinarParticipant;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
 class WebinarController extends Controller
@@ -14,20 +17,23 @@ class WebinarController extends Controller
     {
         $categories = WebinarCategory::get();
         $webinars = Webinar::get();
+        $mentors = Mentor::get();
 
         return view('webinar.index', compact(
-            'categories', 'webinars',
+            'categories', 'webinars', 'mentors'
         ));
     }
 
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'category_id' => 'required',
-            'item' => 'required',
-            'price' => 'required',
-            'quantity' => 'required',
-            'amount' => 'required'
+            'mentors_id' => 'required',
+            'categories_id' => 'required',
+            'title' => 'required',
+            'thumbnail' => 'required|mimes:jpg,jpeg,png,svg',
+            'class_type' => 'required',
+            'date' => 'required',
+            'quota' => 'required'
         ]);
 
         if($validator->fails()){
@@ -36,21 +42,33 @@ class WebinarController extends Controller
                 'data' => $validator->errors()
             ], 400);
         }
+
+        $thumbnail = ($request->thumbnail)
+        ? $request->file('thumbnail')->store("/public/webinar")
+        : null;
         
-        if ($request->category_id) {
-            $categories = explode(',', $request->category_id);
+        if ($request->mentors_id) {
+            $mentors = explode(',', $request->mentors_id);
+        }
+        
+        if ($request->categories_id) {
+            $categories = explode(',', $request->categories_id);
         }
 
-        $data = new Expense([
-            'item' => $request->item,
-            'price' => $request->price,
-            'quantity' => $request->quantity,
-            'amount' => $request->amount,
-            'description' => $request->description
+        $data = new Webinar([
+            'title' => $request->title,
+            'thumbnail' => $thumbnail,
+            'class_type' => $request->class_type,
+            'date' => $request->date,
+            'quota' => $request->quota,
+            'description' => $request->description,
         ]);
 
         if($data->save()){
-            if ($request->category_id) {
+            if ($request->mentors_id) {
+                $data->mentors()->sync($mentors);
+            }
+            if ($request->categories_id) {
                 $data->categories()->sync($categories);
             }
             return response()->json([
@@ -66,13 +84,14 @@ class WebinarController extends Controller
     
     public function detail($id)
     {
-        $expense = Webinar::with('categories')->find($id);
+        $webinar = Webinar::with('mentors')->find($id);
 
-        if($expense){
-            $expense->categoriesSelected = $expense->categories->pluck('id')->toArray();
+        if($webinar){
+            $webinar->mentorsSelected = $webinar->mentors->pluck('id')->toArray();
+            $webinar->categoriesSelected = $webinar->categories->pluck('id')->toArray();
             return response()->json([
                 'status' => true,
-                'data' => $expense
+                'data' => $webinar
             ]);
         }else{
             return response()->json([
@@ -84,12 +103,13 @@ class WebinarController extends Controller
     public function update(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'id' => 'required',
-            'category_id' => 'required',
-            'item' => 'required',
-            'price' => 'required',
-            'quantity' => 'required',
-            'amount' => 'required'
+            'mentors_id' => 'required',
+            'categories_id' => 'required',
+            'title' => 'required',
+            'thumbnail' => 'nullable|mimes:jpg,jpeg,png,svg',
+            'class_type' => 'required',
+            'date' => 'required',
+            'quota' => 'required'
         ]);
 
         if($validator->fails()){
@@ -101,18 +121,30 @@ class WebinarController extends Controller
 
         $data = Webinar::find($request->id);
 
-        if ($request->category_id) {
-            $categories = explode(',', $request->category_id);
+        $thumbnail = ($request->thumbnail)
+        ? $request->file('thumbnail')->store("/public/webinar")
+        : $data->thumbnail;
+
+        if ($request->categories_id) {
+            $categories = explode(',', $request->categories_id);
         }
 
-        $data->item = $request->item ?: $data->item;
-        $data->price = $request->price ?: $data->price;
-        $data->quantity = $request->quantity ?: $data->quantity;
-        $data->amount = $request->amount ?: $data->amount;
+        if ($request->mentors_id) {
+            $mentors = explode(',', $request->mentors_id);
+        }
+
+        $data->title = $request->title ?: $data->title;
+        $data->quota = $request->quota ?: $data->quota;
         $data->description = $request->description ?: $data->description;
+        $data->class_type = $request->class_type ?: $data->class_type;
+        $data->thumbnail = $thumbnail ?: $data->thumbnail;
+        $data->date = $request->date ?: $data->date;
 
         if($data->save()){
-            if ($request->category_id) {
+            if ($request->mentors_id) {
+                $data->mentors()->sync($mentors);
+            }
+            if ($request->categories_id) {
                 $data->categories()->sync($categories);
             }
             return response()->json([
@@ -130,16 +162,23 @@ class WebinarController extends Controller
     {
         $data = Webinar::find($id);
 
-        if($data->delete()){
-            return response()->json([
-                'status' => true,
-                'data' => $data
-            ]);
-        }else{
-            return response()->json([
-                'status' => false
-            ], 404);
+        if($data){
+            if(Storage::exists($data->thumbnail)){
+                Storage::delete($data->thumbnail);
+            }
+
+            if($data->delete()){
+                return response()->json([
+                    'status' => true,
+                    'data' => $data
+                ]);
+            }else{
+                return response()->json([
+                    'status' => false
+                ], 404);
+            }
         }
+
     }
 
     public function categoriesIndex()
@@ -218,7 +257,8 @@ class WebinarController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'id' => 'required',
-            'name' => 'required'
+            'name' => 'required',
+            'slug' => 'required'
         ]);
 
         if($validator->fails()){
@@ -231,6 +271,7 @@ class WebinarController extends Controller
         $data = WebinarCategory::find($request->id);
 
         $data->name = $request->name ?: $data->name;
+        $data->slug = $request->slug ?: $data->slug;
 
         if($data->save()){
             return response()->json([
@@ -247,6 +288,116 @@ class WebinarController extends Controller
     public function categoriesDelete($id)
     {
         $data = WebinarCategory::find($id);
+
+        if($data->delete()){
+            return response()->json([
+                'status' => true,
+                'data' => $data
+            ]);
+        }else{
+            return response()->json([
+                'status' => false
+            ], 404);
+        }
+    }
+
+    public function participantsIndex()
+    {
+        $participants = WebinarParticipant::get();
+        $webinars = Webinar::selectRaw('webinars.*, 
+            ROUND((COUNT(webinar_has_participants.id) / webinars.quota) * 100) AS quota_percentage')
+        ->leftJoin('webinar_has_participants', 'webinars.id', '=', 'webinar_has_participants.webinar_id')
+        ->groupBy('webinars.id')
+        ->havingRaw('COUNT(webinar_has_participants.id) < webinars.quota')
+        ->get();
+
+        return view('webinar.participant.index', compact(
+            'participants', 'webinars'
+        ));
+    }
+
+    public function participantsStore(Request $request)
+    {
+        $validator = Validator::make($request->all(), [
+            'name' => 'required',
+            'webinar_id' => 'required'
+        ]);
+
+        if($validator->fails()){
+            return response()->json([
+                'status' => false,
+                'data' => $validator->errors()
+            ], 400);
+        }
+
+        $data = new WebinarParticipant([
+            'name' => $request->name,
+            'webinar_id' => $request->webinar_id
+        ]);
+
+        if($data->save()){
+            return response()->json([
+                'status' => true,
+                'data' => $data
+            ]);
+        }else{
+            return response()->json([
+                'status' => false
+            ], 404);
+        }
+    }
+
+    public function participantsDetail($id)
+    {
+        $participants = WebinarParticipant::with('webinar')->find($id);
+
+        if($participants){
+            return response()->json([
+                'status' => true,
+                'data' => $participants
+            ]);
+        }else{
+            return response()->json([
+                'status' => false
+            ], 404);
+        }
+    }
+
+    public function participantsUpdate(Request $request)
+    {
+        $validator = Validator::make($request->all(), [
+            'id' => 'required',
+            'name' => 'required',
+            'webinar_id' => 'required'
+        ]);
+
+        if($validator->fails()){
+            return response()->json([
+                'status' => false,
+                'data' => $validator->errors()
+            ], 400);
+        }
+
+        $data = WebinarParticipant::find($request->id);
+
+        $data->name = $request->name ?: $data->name;
+        $data->webinar_id = $request->webinar_id ?: $data->webinar_id;
+
+        if($data->save()){
+            return response()->json([
+                'status' => true,
+                'data' => $data
+            ]);
+        }else{
+            return response()->json([
+                'status' => false
+            ], 404);
+        }
+    }
+
+    public function participantsDelete($id)
+    {
+        $data = WebinarParticipant::find($id);
 
         if($data->delete()){
             return response()->json([
